New  Animation and new enemy model
enemy model and player animation
